requirements.txt
aiogram==3.7.0
openai==2.6.1
ffmpeg-python
python-dotenv
